+++ NoteNest/Models/AppTheme.cs
@@ -1,0 +1,3 @@
+namespace NoteNest.Models;
+
+public enum AppTheme { Light, Dark }